Refactored/requirements.txt
ftfy
openpyxl
Pillow 
psutil 
scipy 
six
typing_extensions
packaging
sentence_transformers
pytorch_lightning
scikit-learn
transformers
pandas
accelerate>=0.26.0
datasets==2.19.1
imblearn
xlrd
tensorboard
